+++ src/cmakex/cmakex-types.h
@@ -179,6 +179,7 @@ struct pkg_request_t : pkg_desc_t
 
     bool git_shallow = false;  // if false, clone only the requested branch at depth=1
     bool define_only = false;
+    bool git_tag_override = false;
 
 private:
 };