+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v1.0, since 2016-10-12
 ----------------------
 
+- Added `GIT_TAG_OVERRIDE` option for `add_pkg`/`def_pkg` commands
 - Added using `<exe-dir>/default-cmakex-presets.yaml` if CMAKEX_PRESET_FILE
   is not set
 - Added `--deps-source`, `--deps-build` and `--deps-install` command-line